added indeterminate possibility
This rewrites the main code of the onDraw() method. So I did loads of
tests on my devices and it seems to work fine. Also there aren’t
annoying graphic errors anymore, because I’m now working with one
single path. Also this is already embedded into the example application
as well.

This is the first major feature for the 1.6.0 version.